Reduce redundancy in plugins module
