+++ mbed_host_tests/host_tests_runner/mbed_base.py
@@ -22,7 +22,7 @@ import os
 import mbed_lstools
 from time import sleep
 from mbed_host_tests import DEFAULT_BAUD_RATE
-import mbed_host_tests.host_tests_plugins as ht_plugins
+import mbed_host_tests.plugins as ht_plugins
 from mbed_host_tests.host_tests_logger import HtrunLogger
