+++ mbed_host_tests/plugins/copy_silabs.py
@@ -17,16 +17,17 @@ limitations under the License.
 Author: Przemyslaw Wirkus <Przemyslaw.Wirkus@arm.com>
 """
 
-from .host_test_plugins import HostTestPluginBase
+import os
+from .base import HostTestPluginBase
 
 
-class HostTestPluginResetMethod_SiLabs(HostTestPluginBase):
+class HostTestPluginCopyMethod_Silabs(HostTestPluginBase):
 
     # Plugin interface
-    name = 'HostTestPluginResetMethod_SiLabs'
-    type = 'ResetMethod'
+    name = 'HostTestPluginCopyMethod_Silabs'
+    type = 'CopyMethod'
     capabilities = ['eACommander', 'eACommander-usb']
-    required_parameters = ['disk']
+    required_parameters = ['image_path', 'destination_disk']
     stable = True
 
     def __init__(self):
@@ -37,7 +38,6 @@ class HostTestPluginResetMethod_SiLabs(HostTestPluginBase):
     def setup(self, *args, **kwargs):
         """ Configure plugin, this function should be called before plugin execute() method is used.
         """
-        # Note you need to have eACommander.exe on your system path!
         self.EACOMMANDER_CMD = 'eACommander.exe'
         return True
 
@@ -54,21 +54,18 @@ class HostTestPluginResetMethod_SiLabs(HostTestPluginBase):
         """
         result = False
         if self.check_parameters(capability, *args, **kwargs) is True:
-            disk = kwargs['disk'].rstrip('/\\')
-
+            image_path = os.path.normpath(kwargs['image_path'])
+            destination_disk = os.path.normpath(kwargs['destination_disk'])
             if capability == 'eACommander':
-                # For this copy method 'disk' will be 'serialno' for eACommander command line parameters
-                # Note: Commands are executed in the order they are specified on the command line
                 cmd = [self.EACOMMANDER_CMD,
-                       '--serialno', disk,
-                       '--resettype', '2', '--reset',]
+                       '--serialno', destination_disk,
+                       '--flash', image_path,
+                       '--resettype', '2', '--reset']
                 result = self.run_command(cmd)
             elif capability == 'eACommander-usb':
-                # For this copy method 'disk' will be 'usb address' for eACommander command line parameters
-                # Note: Commands are executed in the order they are specified on the command line
                 cmd = [self.EACOMMANDER_CMD,
-                       '--usb', disk,
-                       '--resettype', '2', '--reset',]
+                       '--usb', destination_disk,
+                       '--flash', image_path]
                 result = self.run_command(cmd)
         return result
 
@@ -76,4 +73,4 @@ class HostTestPluginResetMethod_SiLabs(HostTestPluginBase):
 def load_plugin():
     """ Returns plugin available in this module
     """
-    return HostTestPluginResetMethod_SiLabs()
+    return HostTestPluginCopyMethod_Silabs()
