+++ mbed_host_tests/plugins/copy_ublox.py
@@ -18,21 +18,16 @@ Author: Przemyslaw Wirkus <Przemyslaw.Wirkus@arm.com>
 """
 
 import os
-from .host_test_plugins import HostTestPluginBase
+from .base import HostTestPluginBase
 
 
-class HostTestPluginCopyMethod_JN51xx(HostTestPluginBase):
+class HostTestPluginCopyMethod_ublox(HostTestPluginBase):
 
     # Plugin interface
-    name = 'HostTestPluginCopyMethod_JN51xx'
+    name = 'HostTestPluginCopyMethod_ublox'
     type = 'CopyMethod'
-    capabilities = ['jn51xx']
-    required_parameters = ['image_path', 'serial']
-
-    def __init__(self):
-        """ ctor
-        """
-        HostTestPluginBase.__init__(self)
+    capabilities = ['ublox']
+    required_parameters = ['image_path']
 
     def is_os_supported(self, os_name=None):
         """! In this implementation this plugin only is supporeted under Windows machines
@@ -49,7 +44,7 @@ class HostTestPluginCopyMethod_JN51xx(HostTestPluginBase):
     def setup(self, *args, **kwargs):
         """! Configure plugin, this function should be called before plugin execute() method is used.
         """
-        self.JN51XX_PROGRAMMER = 'JN51xxProgrammer.exe'
+        self.FLASH_ERASE = 'FlashErase.exe'
         return True
 
     def execute(self, capability, *args, **kwargs):
@@ -58,35 +53,31 @@ class HostTestPluginCopyMethod_JN51xx(HostTestPluginBase):
         @param capability Capability name
         @param args Additional arguments
         @param kwargs Additional arguments
+
         @details Each capability e.g. may directly just call some command line program or execute building pythonic function
+
         @return Capability call return value
         """
-        if not kwargs['image_path']:
-            self.print_plugin_error("Error: image path not specified")
-            return False
-
-        if not kwargs['serial']:
-            self.print_plugin_error("Error: serial port not set (not opened?)")
-            return False
-
         result = False
-        if self.check_parameters(capability, *args, **kwargs):
-            if kwargs['image_path'] and kwargs['serial']:
-                image_path = os.path.normpath(kwargs['image_path'])
-                serial_port = kwargs['serial']
-                if capability == 'jn51xx':
-                    # Example:
-                    # JN51xxProgrammer.exe -s COM15 -f <file> -V0
-                    cmd = [self.JN51XX_PROGRAMMER,
-                           '-s', serial_port,
-                           '-f', image_path,
-                           '-V0'
-                          ]
-                    result = self.run_command(cmd)
+        if self.check_parameters(capability, *args, **kwargs) is True:
+            image_path = os.path.normpath(kwargs['image_path'])
+            if capability == 'ublox':
+                # Example:
+                # FLASH_ERASE -c 2 -s 0xD7000 -l 0x20000 -f "binary_file.bin"
+                cmd = [self.FLASH_ERASE,
+                       '-c',
+                       'A',
+                       '-s',
+                       '0xD7000',
+                       '-l',
+                       '0x20000',
+                       '-f', image_path
+                       ]
+                result = self.run_command(cmd)
         return result
 
 
 def load_plugin():
     """ Returns plugin available in this module
     """
-    return HostTestPluginCopyMethod_JN51xx()
+    return HostTestPluginCopyMethod_ublox()
