+++ mbed_host_tests/plugins/registry.py
