+++ mbed_host_tests/plugins/reset_jn51xx.py
@@ -17,16 +17,16 @@ limitations under the License.
 Author: Przemyslaw Wirkus <Przemyslaw.Wirkus@arm.com>
 """
 
-from .host_test_plugins import HostTestPluginBase
+from .base import HostTestPluginBase
 
 
-class HostTestPluginResetMethod_Stlink(HostTestPluginBase):
+class HostTestPluginResetMethod_JN51xx(HostTestPluginBase):
 
     # Plugin interface
-    name = 'HostTestPluginResetMethod_Stlink'
+    name = 'HostTestPluginResetMethod_JN51xx'
     type = 'ResetMethod'
-    capabilities = ['stlink']
-    required_parameters = []
+    capabilities = ['jn51xx']
+    required_parameters = ['serial']
     stable = False
 
     def __init__(self):
@@ -50,7 +50,7 @@ class HostTestPluginResetMethod_Stlink(HostTestPluginBase):
         """! Configure plugin, this function should be called before plugin execute() method is used.
         """
         # Note you need to have eACommander.exe on your system path!
-        self.ST_LINK_CLI = 'ST-LINK_CLI.exe'
+        self.JN51XX_PROGRAMMER = 'JN51xxProgrammer.exe'
         return True
 
     def execute(self, capability, *args, **kwargs):
@@ -59,23 +59,33 @@ class HostTestPluginResetMethod_Stlink(HostTestPluginBase):
         @param capability Capability name
         @param args Additional arguments
         @param kwargs Additional arguments
-
         @details Each capability e.g. may directly just call some command line program or execute building pythonic function
-
         @return Capability call return value
         """
+        if not kwargs['serial']:
+            self.print_plugin_error("Error: serial port not set (not opened?)")
+            return False
+
         result = False
-        if self.check_parameters(capability, *args, **kwargs) is True:
-            if capability == 'stlink':
-                # Example:
-                # ST-LINK_CLI.exe -Rst -Run
-                cmd = [self.ST_LINK_CLI,
-                       '-Rst', '-Run']
-                result = self.run_command(cmd)
+        if self.check_parameters(capability, *args, **kwargs):
+            if kwargs['serial']:
+                if capability == 'jn51xx':
+                    # Example:
+                    # The device should be automatically reset before the programmer disconnects.
+                    # Issuing a command with no file to program or read will put the device into 
+                    # programming mode and then reset it. E.g.
+                    # $ JN51xxProgrammer.exe -s COM5 -V0
+                    # COM5: Detected JN5179 with MAC address 00:15:8D:00:01:24:E0:37
+                    serial_port = kwargs['serial']
+                    cmd = [self.JN51XX_PROGRAMMER,
+                           '-s', serial_port,
+                           '-V0'
+                          ]
+                    result = self.run_command(cmd)
         return result
 
 
 def load_plugin():
     """ Returns plugin available in this module
     """
-    return HostTestPluginResetMethod_Stlink()
+    return HostTestPluginResetMethod_JN51xx()
